Refactoring: Make all util functions public Use same text printing functionality in HtmlComment as in PrintTextStatement
