+++ src/util/index.js
@@ -1,11 +1,9 @@
-const util = require("./util.js");
 const pluginUtil = require("./pluginUtil.js");
 const publicSymbols = require("./publicSymbols.js");
 const publicFunctions = require("./publicFunctions.js");
 
 const combinedExports = Object.assign(
     {},
-    util,
     pluginUtil,
     publicSymbols,
     publicFunctions
