+++ tests/Failing/controversial.melody.twig
@@ -26,3 +26,8 @@
         entirePlace: isEntirePlace ? entirePlaceData.accommodation.id.id
     } %}
 </article>
+
+<!-- Parentheses stripped. Cannot fix this. Either always parentheses, or never when args empty. -->
+{% include './partials/arrowBtn.melody.twig' with {
+    ref: prev | default(),
+} only %}
