+++ tests/Failing/failing.melody.twig
@@ -1,8 +1,3 @@
-<!-- Cannot fix this. Either always parentheses, or never when args empty -->
-{% include './partials/arrowBtn.melody.twig' with {
-    ref: prev | parensStripped(),
-} only %}
-
 <p class="alert__message">
     {{
         'results_for'
